GenericByteType: change compareTo to use Byte.compare()
This changes the behavior of ByteType.compareTo to return
values different from -1, 0, 1. Exiting code that expects the return
value of compareTo to be either -1, 0 or 1 will break.

+++ src/main/java/net/imglib2/type/numeric/integer/GenericByteType.java
@@ -260,8 +260,7 @@ public abstract class GenericByteType< T extends GenericByteType< T > > extends 
 	@Override
 	public int compareTo( final T other )
 	{
-		// NB: Use Integer.compare because Byte.compare returns values different from -1, 0, 1.
-		return Integer.compare( getByte(), other.getByte() );
+		return Byte.compare( getByte(), other.getByte() );
 	}
 
 	@Override
